Support printing differnt tables for every mesh resource
