+++ emctl/cmd/client/command/apply/apply.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/megaease/easemeshctl/cmd/client/command/flags"
 	"github.com/megaease/easemeshctl/cmd/client/command/meshclient"
-	"github.com/megaease/easemeshctl/cmd/client/resource"
+	"github.com/megaease/easemeshctl/cmd/client/resource/meta"
 	"github.com/megaease/easemeshctl/cmd/client/util"
 	"github.com/megaease/easemeshctl/cmd/common"
 
@@ -53,7 +53,7 @@ func Run(cmd *cobra.Command, flag *flags.Apply) {
 
 	var errs []error
 	for _, vs := range vss {
-		err := vs.Visit(func(mo resource.MeshObject, e error) error {
+		err := vs.Visit(func(mo meta.MeshObject, e error) error {
 			if e != nil {
 				return errors.Wrap(e, "visit failed")
 			}
