+++ emctl/cmd/client/command/get/getter.go
@@ -23,11 +23,12 @@ import (
 
 	"github.com/megaease/easemeshctl/cmd/client/command/meshclient"
 	"github.com/megaease/easemeshctl/cmd/client/resource"
+	"github.com/megaease/easemeshctl/cmd/client/resource/meta"
 	"github.com/megaease/easemeshctl/cmd/common"
 )
 
 // WrapGetterByMeshObject wraps getter for mesh object.
-func WrapGetterByMeshObject(object resource.MeshObject,
+func WrapGetterByMeshObject(object meta.MeshObject,
 	client meshclient.MeshClient, timeout time.Duration) Getter {
 
 	base := baseGetter{
@@ -68,7 +69,7 @@ func WrapGetterByMeshObject(object resource.MeshObject,
 type (
 	// Getter is the getter interface fo mesh object.
 	Getter interface {
-		Get() ([]resource.MeshObject, error)
+		Get() ([]meta.MeshObject, error)
 	}
 
 	baseGetter struct {
@@ -82,7 +83,7 @@ type meshControllerGetter struct {
 	object *resource.MeshController
 }
 
-func (s *meshControllerGetter) Get() ([]resource.MeshObject, error) {
+func (s *meshControllerGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), s.timeout)
 	defer cancelFunc()
 
@@ -92,7 +93,7 @@ func (s *meshControllerGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{meshController}, nil
+		return []meta.MeshObject{meshController}, nil
 	}
 
 	meshControllers, err := s.client.V1Alpha1().MeshController().List(ctx)
@@ -100,7 +101,7 @@ func (s *meshControllerGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(meshControllers))
+	objects := make([]meta.MeshObject, len(meshControllers))
 	for i := range meshControllers {
 		objects[i] = meshControllers[i]
 	}
@@ -113,7 +114,7 @@ type serviceGetter struct {
 	object *resource.Service
 }
 
-func (s *serviceGetter) Get() ([]resource.MeshObject, error) {
+func (s *serviceGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), s.timeout)
 	defer cancelFunc()
 
@@ -123,7 +124,7 @@ func (s *serviceGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{service}, nil
+		return []meta.MeshObject{service}, nil
 	}
 
 	services, err := s.client.V1Alpha1().Service().List(ctx)
@@ -131,7 +132,7 @@ func (s *serviceGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(services))
+	objects := make([]meta.MeshObject, len(services))
 	for i := range services {
 		objects[i] = services[i]
 	}
@@ -144,7 +145,7 @@ type serviceInstanceGetter struct {
 	object *resource.ServiceInstance
 }
 
-func (s *serviceInstanceGetter) Get() ([]resource.MeshObject, error) {
+func (s *serviceInstanceGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), s.timeout)
 	defer cancelFunc()
 
@@ -160,7 +161,7 @@ func (s *serviceInstanceGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{serviceInstance}, nil
+		return []meta.MeshObject{serviceInstance}, nil
 	}
 
 	serviceInstances, err := s.client.V1Alpha1().ServiceInstance().List(ctx)
@@ -168,7 +169,7 @@ func (s *serviceInstanceGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(serviceInstances))
+	objects := make([]meta.MeshObject, len(serviceInstances))
 	for i := range serviceInstances {
 		objects[i] = serviceInstances[i]
 	}
@@ -181,7 +182,7 @@ type canaryGetter struct {
 	object *resource.Canary
 }
 
-func (c *canaryGetter) Get() ([]resource.MeshObject, error) {
+func (c *canaryGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), c.timeout)
 	defer cancelFunc()
 
@@ -191,7 +192,7 @@ func (c *canaryGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{canary}, nil
+		return []meta.MeshObject{canary}, nil
 	}
 
 	canaries, err := c.client.V1Alpha1().Canary().List(ctx)
@@ -199,7 +200,7 @@ func (c *canaryGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(canaries))
+	objects := make([]meta.MeshObject, len(canaries))
 	for i := range canaries {
 		objects[i] = canaries[i]
 	}
@@ -212,7 +213,7 @@ type observabilityTracingsGetter struct {
 	object *resource.ObservabilityTracings
 }
 
-func (o *observabilityTracingsGetter) Get() ([]resource.MeshObject, error) {
+func (o *observabilityTracingsGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), o.timeout)
 	defer cancelFunc()
 
@@ -222,7 +223,7 @@ func (o *observabilityTracingsGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{tracings}, nil
+		return []meta.MeshObject{tracings}, nil
 	}
 
 	tracings, err := o.client.V1Alpha1().ObservabilityTracings().List(ctx)
@@ -230,7 +231,7 @@ func (o *observabilityTracingsGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(tracings))
+	objects := make([]meta.MeshObject, len(tracings))
 	for i := range tracings {
 		objects[i] = tracings[i]
 	}
@@ -243,7 +244,7 @@ type observabilityMetricsGetter struct {
 	object *resource.ObservabilityMetrics
 }
 
-func (o *observabilityMetricsGetter) Get() ([]resource.MeshObject, error) {
+func (o *observabilityMetricsGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), o.timeout)
 	defer cancelFunc()
 
@@ -253,7 +254,7 @@ func (o *observabilityMetricsGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{metrics}, nil
+		return []meta.MeshObject{metrics}, nil
 	}
 
 	metrics, err := o.client.V1Alpha1().ObservabilityMetrics().List(ctx)
@@ -261,7 +262,7 @@ func (o *observabilityMetricsGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(metrics))
+	objects := make([]meta.MeshObject, len(metrics))
 	for i := range metrics {
 		objects[i] = metrics[i]
 	}
@@ -274,7 +275,7 @@ type observabilityOutputServerGetter struct {
 	object *resource.ObservabilityOutputServer
 }
 
-func (o *observabilityOutputServerGetter) Get() ([]resource.MeshObject, error) {
+func (o *observabilityOutputServerGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), o.timeout)
 	defer cancelFunc()
 
@@ -284,7 +285,7 @@ func (o *observabilityOutputServerGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{server}, nil
+		return []meta.MeshObject{server}, nil
 	}
 
 	servers, err := o.client.V1Alpha1().ObservabilityOutputServer().List(ctx)
@@ -292,7 +293,7 @@ func (o *observabilityOutputServerGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(servers))
+	objects := make([]meta.MeshObject, len(servers))
 	for i := range servers {
 		objects[i] = servers[i]
 	}
@@ -305,7 +306,7 @@ type loadBalanceGetter struct {
 	object *resource.LoadBalance
 }
 
-func (l *loadBalanceGetter) Get() ([]resource.MeshObject, error) {
+func (l *loadBalanceGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), l.timeout)
 	defer cancelFunc()
 
@@ -315,7 +316,7 @@ func (l *loadBalanceGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{lb}, nil
+		return []meta.MeshObject{lb}, nil
 	}
 
 	lbs, err := l.client.V1Alpha1().LoadBalance().List(ctx)
@@ -323,7 +324,7 @@ func (l *loadBalanceGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(lbs))
+	objects := make([]meta.MeshObject, len(lbs))
 	for i := range lbs {
 		objects[i] = lbs[i]
 	}
@@ -336,7 +337,7 @@ type tenantGetter struct {
 	object *resource.Tenant
 }
 
-func (t *tenantGetter) Get() ([]resource.MeshObject, error) {
+func (t *tenantGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), t.timeout)
 	defer cancelFunc()
 
@@ -346,7 +347,7 @@ func (t *tenantGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{tenant}, nil
+		return []meta.MeshObject{tenant}, nil
 	}
 
 	tenants, err := t.client.V1Alpha1().Tenant().List(ctx)
@@ -354,7 +355,7 @@ func (t *tenantGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(tenants))
+	objects := make([]meta.MeshObject, len(tenants))
 	for i := range tenants {
 		objects[i] = tenants[i]
 	}
@@ -367,7 +368,7 @@ type resilienceGetter struct {
 	object *resource.Resilience
 }
 
-func (r *resilienceGetter) Get() ([]resource.MeshObject, error) {
+func (r *resilienceGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), r.timeout)
 	defer cancelFunc()
 
@@ -377,7 +378,7 @@ func (r *resilienceGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{resilience}, nil
+		return []meta.MeshObject{resilience}, nil
 	}
 
 	resiliences, err := r.client.V1Alpha1().Resilience().List(ctx)
@@ -385,7 +386,7 @@ func (r *resilienceGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(resiliences))
+	objects := make([]meta.MeshObject, len(resiliences))
 	for i := range resiliences {
 		objects[i] = resiliences[i]
 	}
@@ -398,7 +399,7 @@ type ingressGetter struct {
 	object *resource.Ingress
 }
 
-func (i *ingressGetter) Get() ([]resource.MeshObject, error) {
+func (i *ingressGetter) Get() ([]meta.MeshObject, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), i.timeout)
 	defer cancelFunc()
 
@@ -408,7 +409,7 @@ func (i *ingressGetter) Get() ([]resource.MeshObject, error) {
 			return nil, err
 		}
 
-		return []resource.MeshObject{ingress}, nil
+		return []meta.MeshObject{ingress}, nil
 	}
 
 	ingresses, err := i.client.V1Alpha1().Ingress().List(ctx)
@@ -416,7 +417,7 @@ func (i *ingressGetter) Get() ([]resource.MeshObject, error) {
 		return nil, err
 	}
 
-	objects := make([]resource.MeshObject, len(ingresses))
+	objects := make([]meta.MeshObject, len(ingresses))
 	for i := range ingresses {
 		objects[i] = ingresses[i]
 	}
