+++ emctl/cmd/client/resource/objectcreator.go
@@ -18,14 +18,15 @@
 package resource
 
 import (
+	"github.com/megaease/easemeshctl/cmd/client/resource/meta"
 	"github.com/pkg/errors"
 )
 
 type (
 	// ObjectCreator create a MeshObject
 	ObjectCreator interface {
-		NewFromKind(VersionKind) (MeshObject, error)
-		NewFromResource(MeshResource) (MeshObject, error)
+		NewFromKind(meta.VersionKind) (meta.MeshObject, error)
+		NewFromResource(meta.MeshResource) (meta.MeshObject, error)
 	}
 
 	objectCreator struct{}
@@ -36,15 +37,15 @@ func NewObjectCreator() ObjectCreator {
 	return &objectCreator{}
 }
 
-func (oc *objectCreator) NewFromKind(kind VersionKind) (MeshObject, error) {
-	return oc.new(kind, MetaData{})
+func (oc *objectCreator) NewFromKind(kind meta.VersionKind) (meta.MeshObject, error) {
+	return oc.new(kind, meta.MetaData{})
 }
 
-func (oc *objectCreator) NewFromResource(resource MeshResource) (MeshObject, error) {
+func (oc *objectCreator) NewFromResource(resource meta.MeshResource) (meta.MeshObject, error) {
 	return oc.new(resource.VersionKind, resource.MetaData)
 }
 
-func (oc *objectCreator) new(kind VersionKind, metaData MetaData) (MeshObject, error) {
+func (oc *objectCreator) new(kind meta.VersionKind, metaData meta.MetaData) (meta.MeshObject, error) {
 	apiVersion := kind.APIVersion
 	if apiVersion == "" {
 		apiVersion = DefaultAPIVersion
@@ -101,68 +102,68 @@ func (oc *objectCreator) new(kind VersionKind, metaData MetaData) (MeshObject, e
 }
 
 // NewMeshControllerResource returns a MeshResouce with the mesh controller kind.
-func NewMeshControllerResource(apiVersion, name string) MeshResource {
+func NewMeshControllerResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindMeshController, name)
 }
 
 // NewIngressResource returns a MeshResource with the ingress kind
-func NewIngressResource(apiVersion, name string) MeshResource {
+func NewIngressResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindIngress, name)
 }
 
 // NewServiceResource returns a MeshResource with the service kind
-func NewServiceResource(apiVersion, name string) MeshResource {
+func NewServiceResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindService, name)
 }
 
 // NewServiceInstanceResource returns a MeshResource with the service kind
-func NewServiceInstanceResource(apiVersion, name string) MeshResource {
+func NewServiceInstanceResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindServiceInstance, name)
 }
 
 // NewCanaryResource returns a MeshResource with the canary kind
-func NewCanaryResource(apiVersion, name string) MeshResource {
+func NewCanaryResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindCanary, name)
 }
 
 // NewLoadBalanceResource returns a MeshResource with the loadbalance kind
-func NewLoadBalanceResource(apiVersion, name string) MeshResource {
+func NewLoadBalanceResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindLoadBalance, name)
 }
 
 // NewResilienceResource returns a MeshResource with the resilience kind
-func NewResilienceResource(apiVersion, name string) MeshResource {
+func NewResilienceResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindResilience, name)
 }
 
 // NewObservabilityTracingsResource returns a MeshResource with the observability tracings kind
-func NewObservabilityTracingsResource(apiVersion, name string) MeshResource {
+func NewObservabilityTracingsResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindObservabilityTracings, name)
 }
 
 // NewObservabilityMetricsResource returns a MeshResource with the observability metrics kind
-func NewObservabilityMetricsResource(apiVersion, name string) MeshResource {
+func NewObservabilityMetricsResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindObservabilityMetrics, name)
 }
 
 // NewObservabilityOutputServerResource returns a MeshResource with the observability output service kind
-func NewObservabilityOutputServerResource(apiVersion, name string) MeshResource {
+func NewObservabilityOutputServerResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindObservabilityOutputServer, name)
 }
 
 // NewTenantResource returns a MeshResource with the tenant kind
-func NewTenantResource(apiVersion, name string) MeshResource {
+func NewTenantResource(apiVersion, name string) meta.MeshResource {
 	return NewMeshResource(apiVersion, KindTenant, name)
 }
 
 // NewMeshResource returns a generic MeshResource
-func NewMeshResource(api, kind, name string) MeshResource {
-	return MeshResource{
-		VersionKind: VersionKind{
+func NewMeshResource(api, kind, name string) meta.MeshResource {
+	return meta.MeshResource{
+		VersionKind: meta.VersionKind{
 			APIVersion: api,
 			Kind:       kind,
 		},
-		MetaData: MetaData{
+		MetaData: meta.MetaData{
 			Name: name,
 		},
 	}
