+++ emctl/cmd/client/util/visitor.go
@@ -29,6 +29,7 @@ import (
 	"time"
 
 	"github.com/megaease/easemeshctl/cmd/client/resource"
+	"github.com/megaease/easemeshctl/cmd/client/resource/meta"
 
 	"github.com/go-resty/resty/v2"
 	"github.com/pkg/errors"
@@ -50,7 +51,7 @@ type Visitor interface {
 }
 
 // VisitorFunc executes visition logic
-type VisitorFunc func(resource.MeshObject, error) error
+type VisitorFunc func(meta.MeshObject, error) error
 
 // RawExtension is a raw struct that holds raw information of the spec
 type RawExtension struct {
@@ -218,7 +219,7 @@ func (v *streamVisitor) Visit(fn VisitorFunc) error {
 	return finalErr
 }
 
-func (v *streamVisitor) decodeMeshObject(data []byte, source string) (resource.MeshObject, error) {
+func (v *streamVisitor) decodeMeshObject(data []byte, source string) (meta.MeshObject, error) {
 	meshObject, _, err := v.Decoder.Decode(data)
 	if err != nil {
 		return nil, err
@@ -323,20 +324,20 @@ func adaptCommandKind(kind string) string {
 }
 
 func (v *commandVisitor) Visit(fn VisitorFunc) error {
-	vk := resource.VersionKind{
+	vk := meta.VersionKind{
 		APIVersion: resource.DefaultAPIVersion,
 		Kind:       v.Kind,
 	}
 
-	var mo resource.MeshObject
+	var mo meta.MeshObject
 	var err error
 
 	if v.Name == "" {
 		mo, err = v.oc.NewFromKind(vk)
 	} else {
-		mo, err = v.oc.NewFromResource(resource.MeshResource{
+		mo, err = v.oc.NewFromResource(meta.MeshResource{
 			VersionKind: vk,
-			MetaData: resource.MetaData{
+			MetaData: meta.MetaData{
 				Name: v.Name,
 			},
 		})
